requirements.txt
portia-sdk-python
requests
streamlit
python-dotenv
pydantic 